Fix: delete 로직에도 댓글작성자 검증 로직 추가

+++ src/apis/comments/comments.controller.ts
@@ -8,12 +8,14 @@ import {
   Post,
   Req,
   UnprocessableEntityException,
+  UseGuards,
 } from '@nestjs/common';
 import { CommentsService } from './comments.service';
 import { Comment } from './entities/comment.entity';
 import { CreateCommentsInput } from './dto/create-comments.input';
 import { Request } from 'express';
 import { IAuthUser } from 'src/commons/interfaces/context';
+import { AuthGuard } from '@nestjs/passport';
 
 @Controller('comments')
 export class CommentsController {
@@ -22,6 +24,7 @@ export class CommentsController {
   ) {}
 
   @Post('/:pubbleId')
+  @UseGuards(AuthGuard('access'))
   createComment(
     @Body('createCommentsInput') createCommentsInput: CreateCommentsInput,
     @Req() req: Request & IAuthUser,
@@ -41,6 +44,7 @@ export class CommentsController {
   }
 
   @Patch('/:id')
+  @UseGuards(AuthGuard('access'))
   updateComment(
     @Param('id') id: string, //
     @Body('content') content: string,
@@ -51,7 +55,12 @@ export class CommentsController {
   }
 
   @Delete('/:id')
-  deleteComment(@Param('id') id: string): Promise<boolean> {
-    return this.commentsService.delete({ id });
+  @UseGuards(AuthGuard('access'))
+  deleteComment(
+    @Param('id') id: string, //
+    @Req() req: Request & IAuthUser,
+  ): Promise<boolean> {
+    if (!req.user) throw new UnprocessableEntityException('unauthorized');
+    return this.commentsService.delete({ id, authorId: req.user.id });
   }
 }
